L23_django_ORM/serp/pyproject.toml
[tool.poetry]
name = "serp"
version = "0.1.0"
description = ""
authors = ["Lev Pikulev <pikulev.l.v@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Django = "^3.1.3"
django-debug-toolbar = "^3.1.1"

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
